Add groupBy alerts

+++ x-pack/legacy/plugins/infra/server/routes/alerts/delete_metric_alert.ts
@@ -4,6 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import Boom from 'boom';
 import { InfraBackendLibs } from '../../lib/infra_types';
 import { internalInfraFrameworkRequest } from '../../lib/adapters/framework/adapter_types';
 import { infraMetricAlertSavedObjectType } from '../../lib/alerting/saved_object_mappings';
@@ -28,7 +29,20 @@ export const initDeleteMetricAlert = ({ framework }: InfraBackendLibs) =>
       const { id } = req.query;
 
       try {
-        const result = await alertsClient.delete({ id });
+        const savedAlert = await savedObjectsClient.get(infraMetricAlertSavedObjectType, id);
+        if (savedAlert.attributes.childAlerts) {
+          for (const alertID of savedAlert.attributes.childAlerts) {
+            await alertsClient.delete({ id: alertID });
+            await savedObjectsClient.delete(infraMetricAlertSavedObjectType, alertID);
+          }
+        } else {
+          if (savedAlert.attributes.childOf) {
+            throw new Boom(
+              `Cannot delete the child of a groupBy alert. You must delete the parent alert, which has the ID of ${savedAlert.attributes.childOf})`
+            );
+          }
+          await alertsClient.delete({ id });
+        }
         await savedObjectsClient.delete(infraMetricAlertSavedObjectType, id);
 
         return res.response().code(204);
